+++ tests/profiles/googlefonts_test.py
@@ -2585,7 +2585,8 @@ def test_check_name_copyright_length():
   assert status == FAIL
 
 
-def test_check_fontdata_namecheck():
+# namecheck website is down. See: https://github.com/googlefonts/fontbakery/issues/2484
+def DISABLED_test_check_fontdata_namecheck():
   """ Familyname is unique according to namecheck.fontdata.com """
   from fontbakery.profiles.googlefonts import (com_google_fonts_check_fontdata_namecheck as check,
                                                      familyname)