+++ CHANGELOG.md
@@ -9,6 +9,9 @@ A more detailed list of changes is available in the corresponding milestones for
 ### Deprecated checks
   - **[com.google.fonts/check/currency_chars]:** we now have a much broader glyph coverage check: com.google.fonts/check/glyph_coverage (issue #2498)
 
+### Bug fixes
+  - The HTML report now actually defaults to "sans-serif" as the body font.
+
 
 ## 0.7.4 (2019-May-06)
 ### Note-worthy code changes