+++ CHANGELOG.md
@@ -11,6 +11,7 @@ A more detailed list of changes is available in the corresponding milestones for
 
 ### Bug fixes
   - The HTML report now actually defaults to "sans-serif" as the body font.
+  - Restrict com.google.fonts/check/repo/dirname_matches_nameid_1 to static fonts as the `fontbakery.util.get_regular` function does not support variable fonts yet. (issue #2509)
 
 
 ## 0.7.4 (2019-May-06)