+++ CHANGELOG.md
@@ -2,7 +2,12 @@ Below are the most important changes from each release.
 A more detailed list of changes is available in the corresponding milestones for each release in the Github issue tracker (https://github.com/googlefonts/fontbakery/milestones?state=closed).
 
 
-## 0.7.5 (2019-May-20)
+## 0.7.6 (2019-Jun-??)
+### Bug fixes
+  - ...
+
+
+## 0.7.5 (2019-May-24)
 ### Note-worthy code changes
   - The conditions from the googlefonts profile were split out into their own separate file